refactor: unify the role selector
