+++ frontend/src/components/v2/Select/RoleSelect.vue
@@ -5,17 +5,24 @@
     :disabled="disabled"
     :clearable="clearable"
     :options="availableRoleOptions"
+    :max-tag-count="'responsive'"
+    :filterable="true"
     :placeholder="$t('settings.members.select-role', multiple ? 2 : 1)"
-    @update:value="$emit('update:value', $event)"
+    @update:value="onValueUpdate"
+  />
+  <FeatureModal
+    feature="bb.feature.custom-role"
+    :open="showFeatureModal"
+    @cancel="showFeatureModal = false"
   />
 </template>
 
 <script setup lang="tsx">
 import type { SelectGroupOption, SelectOption } from "naive-ui";
 import { NSelect } from "naive-ui";
-import { computed } from "vue";
+import { computed, ref } from "vue";
 import { t } from "@/plugins/i18n";
-import { useAppFeature, useRoleStore } from "@/store";
+import { useAppFeature, useRoleStore, featureToRef } from "@/store";
 import {
   PRESET_PROJECT_ROLES,
   PRESET_ROLES,
@@ -31,12 +38,14 @@ const props = withDefaults(
     clearable?: boolean;
     multiple?: boolean;
     suffix?: string;
+    includeWorkspaceRoles?: boolean;
     size?: "tiny" | "small" | "medium" | "large";
   }>(),
   {
     clearable: true,
     value: undefined,
     multiple: false,
+    includeWorkspaceRoles: true,
     suffix: () =>
       ` (${t("common.optional")}, ${t(
         "role.project-roles.apply-to-all-projects"
@@ -45,27 +54,18 @@ const props = withDefaults(
   }
 );
 
-defineEmits<{
+const emit = defineEmits<{
   (event: "update:value", val: string | string[]): void;
 }>();
 
 const roleStore = useRoleStore();
 const hideProjectRoles = useAppFeature("bb.feature.members.hide-project-roles");
+const showFeatureModal = ref(false);
+const hasCustomRoleFeature = featureToRef("bb.feature.custom-role");
 
 const availableRoleOptions = computed(
   (): (SelectOption | SelectGroupOption)[] => {
     const roleGroups = [
-      {
-        type: "group",
-        key: "workspace-roles",
-        label: t("role.workspace-roles.self"),
-        children: PRESET_WORKSPACE_ROLES.filter(
-          (role) => role !== PresetRoleType.WORKSPACE_MEMBER
-        ).map((role) => ({
-          label: displayRoleTitle(role),
-          value: role,
-        })),
-      },
       {
         type: "group",
         key: "project-roles",
@@ -76,6 +76,19 @@ const availableRoleOptions = computed(
         })),
       },
     ];
+    if (props.includeWorkspaceRoles) {
+      roleGroups.unshift({
+        type: "group",
+        key: "workspace-roles",
+        label: t("role.workspace-roles.self"),
+        children: PRESET_WORKSPACE_ROLES.filter(
+          (role) => role !== PresetRoleType.WORKSPACE_MEMBER
+        ).map((role) => ({
+          label: displayRoleTitle(role),
+          value: role,
+        })),
+      });
+    }
     if (hideProjectRoles.value) {
       return roleGroups[0].children;
     }
@@ -96,4 +109,22 @@ const availableRoleOptions = computed(
     return roleGroups;
   }
 );
+
+const includeCustomRole = (values: string[]) => {
+  return values.some((val) => !PRESET_ROLES.includes(val));
+};
+
+const onValueUpdate = (val: string | string[]) => {
+  let hasCustomRole = false;
+  if (Array.isArray(val)) {
+    hasCustomRole = includeCustomRole(val);
+  } else {
+    hasCustomRole = includeCustomRole([val]);
+  }
+  if (hasCustomRole && !hasCustomRoleFeature.value) {
+    showFeatureModal.value = true;
+    return;
+  }
+  emit("update:value", val);
+};
 </script>
