+++ frontend/src/components/v2/Select/index.ts
@@ -4,7 +4,6 @@ import DatabaseSelect from "./DatabaseSelect.vue";
 import EnvironmentSelect from "./EnvironmentSelect.vue";
 import GroupSelect from "./GroupSelect.vue";
 import InstanceSelect from "./InstanceSelect.vue";
-import ProjectRoleSelect from "./ProjectRoleSelect.vue";
 import ProjectSelect from "./ProjectSelect.vue";
 import RoleSelect from "./RoleSelect.vue";
 import UserSelect from "./UserSelect.vue";
@@ -16,7 +15,6 @@ export {
   DatabaseGroupSelect,
   EnvironmentSelect,
   ProjectSelect,
-  ProjectRoleSelect,
   UserSelect,
   GroupSelect,
   RoleSelect,
